requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-google-genai
trustcall
langgraph-cli[inmem]
pandas
sentence-transformers 
faiss-cpu 
numpy 
tiktoken